pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-flask>=1.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
]
